rowan::api

Struct SyntaxToken

pub struct SyntaxToken<L: Language> { /* private fields */ }

Implementations

impl<L: Language> SyntaxToken<L>

pub fn replace_with(&self, new_token: GreenToken) -> GreenNode

Returns a green tree, equal to the green tree this token belongs two, except with this token substitute. The complexity of operation is proportional to the depth of the tree

pub fn kind(&self) -> L::Kind

pub fn text_range(&self) -> TextRange

pub fn index(&self) -> usize

pub fn text(&self) -> &str

pub fn green(&self) -> &GreenTokenData

pub fn parent(&self) -> Option<SyntaxNode<L>>

pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode<L>>

๐Deprecated: use SyntaxToken::parent_ancestors instead

Iterator over all the ancestors of this token excluding itself.

pub fn parent_ancestors(&self) -> impl Iterator<Item = SyntaxNode<L>>

Iterator over all the ancestors of this token excluding itself.

pub fn next_sibling_or_token(&self) -> Option<SyntaxElement<L>>

pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement<L>>

pub fn siblings_with_tokens( &self, direction: Direction, ) -> impl Iterator<Item = SyntaxElement<L>>

pub fn next_token(&self) -> Option<SyntaxToken<L>>

Next token in the tree (i.e, not necessary a sibling).

pub fn prev_token(&self) -> Option<SyntaxToken<L>>

Previous token in the tree (i.e, not necessary a sibling).

pub fn detach(&self)