Module rowan::green

The "green tree" is the immutable, atomically reference counted tree.

Leaf tokens are GreenToken and non-leaf nodes are GreenNode. Both of these types are unsized and are only used behind references. All tree nodes are aggressively reference counted and reused, to minimize space taken up by the immutable tree and to maximize sharing.

Structs

Checkpoint
Checkpoint for maybe wrapping a node. See GreenTreeBuilder::checkpoint.

Children
Children of a node in the immutable green tree.

GreenBuilder
Construction cache for green tree elements.

GreenNode
Internal node in the immutable tree. It has other nodes and tokens as children.

GreenToken
Leaf node in the immutable tree.

GreenTreeBuilder
Builder for a green tree.