Struct rowan::green::GreenBuilder

pub struct GreenBuilder { /* fields omitted */ }

Construction cache for green tree elements.

As the green tree is immutable, identical nodes can be deduplicated. For example, all nodes representing the #[inline] attribute can be deduplicated and refer to the same green node in memory, despite their distribution throughout the source code.

Methods

impl GreenBuilder

pub fn node<I, E>(&mut self, kind: Kind, children: I) -> Arc<GreenNode> where
E: Into<NodeOrToken<Arc<GreenNode>, Arc<GreenToken>>>,
I: IntoIterator<Item = E>,
I::IntoIter: ExactSizeIterator + TrustedLen,

Create a new node or clone a new Arc to an existing equivalent one.

This checks children for identity equivalence, not structural, so it is O(children.len()) and only caches higher-level nodes if the lower-level nodes have also been cached.

pub fn token(&mut self, kind: Kind, text: &str) -> Arc<GreenToken>

Create a new token or clone a new Arc to an existing equivalent one.