Struct rowan::syntax::Token

#[repr(transparent)]
pub struct Token<Lang> { /* fields omitted */ }

Token in the syntax tree.

Methods

impl<Lang: Language> Token<Lang>

Constructors and transformations between Tokens and Nodes.

pub fn new(node: Node<Lang>) -> Option<Token<Lang>>

Create a new token from a node.

This does not check if the language considers the node a token, just whether the node is backed by a green token. If you need the language check, check Node::is_token.

pub fn new_ref(node: &Node<Lang>) -> Option<&Token<Lang>>

Create a new token reference from a node reference.

This does not check if the language considers the node a token, just whether the node is backed by a green token. If you need the language check, use Node::as_token.

pub fn into_node(self) -> Node<Lang>

Convert this token back into a node.

pub fn as_node(&self) -> &Node<Lang>

Convert this token back into a node.

impl<Lang: Language> Token<Lang>

Tree traversal iterators.

pub fn ancestors(self) -> impl Iterator<Item = Node<Lang>>

The parent chain from this token, starting with this node.

pub fn siblings(self, direction: Direction) -> impl Iterator<Item = Node<Lang>>

Sibling trees starting at and including this token.

impl<Lang: Language> Token<Lang>

Accessors.

pub fn green(&self) -> ArcBorrow<GreenToken>

The green token backing this token.

pub fn kind(&self) -> Kind

The kind of this token.

pub fn text(self) -> Text

The text of this token.

pub fn text_range(&self) -> StrRange

The range of text this token covers.

pub fn as_str(&self) -> &str

The raw string of this token.

pub fn replace_with(
&self,
replacement: impl Into<NodeOrToken<Arc<GreenNode>, Arc<GreenToken>>>
) -> NodeOrToken<Arc<GreenNode>, Arc<GreenToken>>

Create a new green tree with this node replaced. The complexity is proportional to the depth of the tree.

pub fn parent(self) -> Option<Node<Lang>>

The parent of this token.

pub fn next_sibling(self) -> Option<Node<Lang>>

The next subtree of this node's parent.

pub fn prev_sibling(self) -> Option<Node<Lang>>

The previous subtree of this node's parent.