Crate rowan

A generic library for lossless syntax trees.

Re-exports

pub use crate::green::GreenNode;

pub use crate::green::GreenToken;

pub use crate::syntax::Language;

pub use crate::syntax::Node;

pub use crate::syntax::Text;

pub use crate::syntax::Token;

pub use rc_borrow::ArcBorrow;

pub use str_index::StrIndex;

pub use str_index::StrRange;

Modules

green
The "green tree" is the immutable, atomically reference counted tree.

syntax
The "syntax tree" is a thread-local view of the green tree with parent pointers and text offsets.

Structs

Kind
Raw kind tag for each token or node in the tree.

Enums

Direction
Direction of an iterator.

NodeOrToken
Convenience enum wrapping either a node or a token.

WalkEvent
Steps in the tree walking process.